1/6/27
Website Updated! The "About" me section now includes a timeline of events that describe who I am and what accomplishments I have achieved over the years. I have removed the "Contact" me panel in favor of a card on the splash page which includes my LinkedIn and Instagram. The news section will most likley have a timeline addition in favor of the current format of "blocks" in the future.