1/7/2022
Website Updated! I removed the "About" me section and moved the timeline into the news section and replaced the "block" style in favor of the timeline. I added promotional material for "Fruitilicious!" in a new page called "Apps". In the future, I plan for this to promote the other apps that are currently being developed but are not yet ready for promotion.