4/1/2022
The Navigation bar was changed to be centered to better flow with the centered design of the elements on the website. The News page was updated to now allow users to view older versions of the website. The Projects page was updated to include Fruitilicious, and downloads were moved offsite to not overflow the size of the website with each update, since website updates are now being archived. The fade-in feature was reintroduced and a loading screen was added when loading is depending on fetching media (Images/Videos).
      Implementation of dynamic webpages is introduced but not optimized for all pages. Finally, a lower navigation bar was added to quickly access my Instagram and LinkedIn and included a copyright notice on the webpages derivate works. April Fools jokes need not apply!